requirements.txt
openai==2.21.0
python-dotenv==1.2.1
google-api-python-client==2.190.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4